staging: zram: Fix potential data corruption
Any source page can compress exactly into the same size as uncompressed
data. With recent commit

9918907 staging: zram: Fix handling of incompressible pages

a page of which compressed size exceeds uncompressed size is being
stored in uncompressed form, then decompressing code distunguishes
such pages by size equals to page size, which can be a compressed page
as well, because of the condition in compressing code.

This patches modifies compressing condition to greater or equal,
so any page that does not compress to less than page size is stored
in uncompressed form.

As a side effect it might also give some minor speedup with read
accesses of uncompressible data stored in zram.

Signed-off-by: Tomasz Figa <tomasz.figa@gmail.com>